Move cycle legion to handle_input
