+++ src/input/hotkey.h
@@ -2,6 +2,7 @@
 #define INPUT_HOTKEY_H
 
 typedef struct {
+    int cycle_legion;
     int show_advisor;
     int show_overlay;
     int toggle_overlay;
